sync

+++ DataLayer/Interface/Entities/IPanelLineNumbers.cs
@@ -5,6 +5,6 @@ namespace DataLayer.Interface.Entities
    public interface IPanelLineNumbers:IHasGuid
     {
         Guid PanelGuid { get; set; }
-        int LineNumber { get; set; }
+        long LineNumber { get; set; }
     }
 }
